Move credential matching code to its own file.
